Feat: Add support for using globs in arrays and functions
This commit introduces support for using glob patterns in arrays
and functions provided to the `components`. This feature lets you
for instance include multiple glob patterns under the same section.
This can be useful if your repository has related components spread
across multiple folders, in which case maintaining a single glob
pattern that would include all folders could get tricky. This feature
may for instance be useful if you work with a monorepo.