WIP: Debug named export error for example components
iss: #1281

Updated examples-loader to import component with named export
Named export of Basic button is not working, but CounterButton is
now falling with an error:

```
Invariant Violation: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: object.
Check the render method of `FunctionComponentWrapper`.

```