Fix section pages if section name has inappropriate symbols
If you have a forward slash in your section name, the pagePerSection links are all broken,
 resulting in "Page not found".
Removing the forward slashes in the names fixes the issue.

For example:
  pagePerSection links break when section name includes a "/", like `name: '@foo/components'`

Solution is just encode name in the getUrl and then correctly
decode it in the router logic.

iss: https://github.com/styleguidist/react-styleguidist/issues/1332