Fix component name capitalization issue
iss: https://github.com/styleguidist/react-styleguidist/issues/1381

It is happening in a case when react-docgen (any propsParser) is failed
and we are trying to guess displayName based on file path.

We capitalized every case in the string instead of just words separated
by `-` started with small letter.

So, by mistake `ButtonTS` in file name becomes `ButtonTs` display name for
component when we expect `ButtonTS`