Fix: Add custom display code to Examples/Playground.
Add a new property called `displayCode` to the `Playground` component.
When this property is set, the code displayed in the editor below the
example preview will come from `displayCode` instead of the usual `code`
property.

This can be used to enable hot-reloading and transpiling during
development, see the relevant pull request description.