Enhancement: Modifier to add whitespace between example components
We want users to be able to easily add padding between example components without
adding manual styling to their example code. This adds the `padded` modifier option
to the Playground, which adds additional styling to the PlaygroundRenderer when
`true` that adds padding between each example component in that block.

Issue #841