Fix: Correct Usage of CleanWebpackPlugin (#1565)
I was still using an option that no longer exists. Fixed to mimic the original behavior. The option `root` was removed in v2, but wasn't changed during the upgrade in https://github.com/styleguidist/react-styleguidist/commit/98042591dc75cad8cdf88cc2a2f0817d69e5e66d

It's deleting everything in the styleguidistDir folder which isn't the way it was before.
After this change it only deletes files in build/.