chore: ignore 'impossible' situation in test coverage
There are almost no situations in which ReactDOM.render throws a synchronous
error. The only one currently is when the provided container is not a DOM
node, but that situation is impossible to cover in a test. It seems fine
to keep the catch clause here for when ReactDOM.render possibly changes in
the future, but we should ignore it w.r.t. test coverage.